refactor: Move PartitionTemplate and friends to their own module
