+++ data_types/src/lib.rs
@@ -19,6 +19,8 @@ mod columns;
 pub use columns::*;
 mod namespace_name;
 pub use namespace_name::*;
+mod partition_template;
+pub use partition_template::*;
 
 use observability_deps::tracing::warn;
 use schema::{sort::SortKey, TIME_COLUMN_NAME};
@@ -846,40 +848,6 @@ impl ChunkOrder {
     pub fn get(&self) -> i64 {
         self.0
     }
-}
-
-/// `PartitionTemplate` is used to compute the partition key of each row that gets written. It can
-/// consist of a column name and its value, a formatted time, or a string column and regex captures
-/// of its value. For columns that do not appear in the input row, a blank value is output.
-///
-/// The key is constructed in order of the template parts; thus ordering changes what partition key
-/// is generated.
-#[derive(Debug, Eq, PartialEq, Clone)]
-#[allow(missing_docs)]
-pub struct PartitionTemplate {
-    pub parts: Vec<TemplatePart>,
-}
-
-impl Default for PartitionTemplate {
-    fn default() -> Self {
-        Self {
-            parts: vec![TemplatePart::TimeFormat("%Y-%m-%d".to_owned())],
-        }
-    }
-}
-
-/// `TemplatePart` specifies what part of a row should be used to compute this
-/// part of a partition key.
-#[derive(Debug, Eq, PartialEq, Clone)]
-pub enum TemplatePart {
-    /// The value in a named column
-    Column(String),
-    /// Applies a  `strftime` format to the "time" column.
-    ///
-    /// For example, a time format of "%Y-%m-%d %H:%M:%S" will produce
-    /// partition key parts such as "2021-03-14 12:25:21" and
-    /// "2021-04-14 12:24:21"
-    TimeFormat(String),
 }
 
 /// Represents a parsed delete predicate for evaluation by the InfluxDB IOx
